1. Move calling scripts to wooMain.main and wooMain.batch, use console_script hooks in setup.py to install those 2. Fix compilation with broken OpenCL c++ headers in Ubuntu 12.10 3. Fix some compilation warnings 4.

+++ pkg/dem/LawTester.cpp
@@ -201,7 +201,7 @@ void LawTester::run(){
 				py::exec(stg->done.c_str(),globals,locals);
 				errCmd=nullptr;
 			}
-			if(stage<stages.size()-1) stage++;
+			if(stage<(int)stages.size()-1) stage++;
 			else{ // finished
 				py::exec(done.c_str(),globals,locals);
 			};
